+++ src/main/java/com/sikoramarek/fiszki/service/QuestionService.java
@@ -26,18 +26,18 @@ import java.util.stream.Collectors;
 @Service
 public class QuestionService {
 
-	private QuestionRepository questionsRepository;
-	private AnswerRepository answersRepository;
+	private QuestionRepository questionRepository;
+	private AnswerRepository answerRepository;
 	private TagRepository tagRepository;
 	private UserRepository userRepository;
 
-	public QuestionService(QuestionRepository questionsRepository,
+	public QuestionService(QuestionRepository questionRepository,
 	                       TagRepository tagRepository,
-	                       AnswerRepository answersRepository,
+	                       AnswerRepository answerRepository,
 	                       UserRepository userRepository) {
 
-		this.questionsRepository = questionsRepository;
-		this.answersRepository = answersRepository;
+		this.questionRepository = questionRepository;
+		this.answerRepository = answerRepository;
 		this.tagRepository = tagRepository;
 		this.userRepository = userRepository;
 	}
@@ -53,13 +53,13 @@ public class QuestionService {
 
 	public Page<Question> getPageableQuestions(int page, int size) {
 		if (checkForAdmin()) {
-			return questionsRepository.findAll(PageRequest.of(page, size));
+			return questionRepository.findAll(PageRequest.of(page, size));
 		}
-		return questionsRepository.findAllByAcceptedTrue(PageRequest.of(page, size));
+		return questionRepository.findAllByAcceptedTrue(PageRequest.of(page, size));
 	}
 
 	public ResponseEntity<List<Question>> getQuestionById(Long question_id) {
-		Optional<Question> optionalQuestion = questionsRepository.findQuestionById(question_id);
+		Optional<Question> optionalQuestion = questionRepository.findQuestionById(question_id);
 		return optionalQuestion
 				.map(question ->
 						new ResponseEntity<>(Collections.singletonList(question), HttpStatus.OK))
@@ -68,14 +68,14 @@ public class QuestionService {
 	}
 
 	public ResponseEntity<Question> newQuestion(Question question) {
-		questionsRepository.save(question);
+		questionRepository.save(question);
 		if (question.getAnswers() != null && question.getAnswers().size() > 0) {
 			try {
 				Collection<Answer> answers = question.getAnswers();
 				answers.forEach(answer -> {
 					answer.setQuestion(question);
 					answer.setUser(question.getUser());
-					answersRepository.save(answer);
+					answerRepository.save(answer);
 				});
 			} catch (DataAccessException | ConstraintViolationException e) {
 				return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -88,7 +88,7 @@ public class QuestionService {
 		if (principal == null) {
 			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 		}
-		Optional<Question> optionalQuestion = questionsRepository.findById(question_id);
+		Optional<Question> optionalQuestion = questionRepository.findById(question_id);
 		if (!optionalQuestion.isPresent()) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
@@ -105,7 +105,7 @@ public class QuestionService {
 		question.setTitle(newQuestion.getTitle());
 		question.setQuestion(newQuestion.getQuestion());
 		try {
-			questionsRepository.save(question);
+			questionRepository.save(question);
 		} catch (DataAccessException | ConstraintViolationException e) {
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		}
@@ -116,7 +116,7 @@ public class QuestionService {
 		if (principal == null) {
 			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 		}
-		Optional<Question> optionalQuestion = questionsRepository.findById(question_id);
+		Optional<Question> optionalQuestion = questionRepository.findById(question_id);
 		if (!optionalQuestion.isPresent()) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
@@ -124,25 +124,25 @@ public class QuestionService {
 		if (!(principal.getName().equals(question.getUser()) || checkForAdmin())) {
 			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 		}
-		questionsRepository.delete(question);
+		questionRepository.delete(question);
 		return new ResponseEntity<>(question, HttpStatus.OK);
 	}
 
 	public ResponseEntity<List<Question>> getRandom(Principal principal) {
 		Page<Question> questionPage;
 		if (principal != null && getCurrentUserKnownQuestionIds().size() > 0) {
-			Long quantity = questionsRepository
+			Long quantity = questionRepository
 					.countQuestionByAcceptedTrueAndIdNotIn(
 							getCurrentUserKnownQuestionIds());
 			int index = (int) (Math.random() * quantity);
-			questionPage = questionsRepository
+			questionPage = questionRepository
 					.findQuestionsByIdNotIn(
 							getCurrentUserKnownQuestionIds(),
 							PageRequest.of(index, 1, Sort.unsorted()));
 		} else {
-			Long quantity = questionsRepository.countByAcceptedTrue();
+			Long quantity = questionRepository.countByAcceptedTrue();
 			int index = (int) (Math.random() * quantity);
-			questionPage = questionsRepository.findAllByAcceptedTrue(PageRequest.of(index, 1, Sort.unsorted()));
+			questionPage = questionRepository.findAllByAcceptedTrue(PageRequest.of(index, 1, Sort.unsorted()));
 		}
 		if (questionPage.hasContent()) {
 			return new ResponseEntity<>(Collections.singletonList(questionPage.getContent().get(0)), HttpStatus.OK);
@@ -160,7 +160,7 @@ public class QuestionService {
 		if (principal != null) {
 			String userName = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 			UserModel user = userRepository.getUserByUsername(userName);
-			Optional<Question> optionalQuestion = questionsRepository.findQuestionById(question_id);
+			Optional<Question> optionalQuestion = questionRepository.findQuestionById(question_id);
 			if (optionalQuestion.isPresent()) {
 				Question question = optionalQuestion.get();
 				Set<UserModel> usersKnownThisQuestion = question.getUsersKnownThisQuestion();
@@ -169,7 +169,7 @@ public class QuestionService {
 				} else {
 					usersKnownThisQuestion.add(user);
 				}
-				questionsRepository.save(question);
+				questionRepository.save(question);
 				return new ResponseEntity(HttpStatus.OK);
 			} else {
 				return new ResponseEntity(HttpStatus.NOT_FOUND);
@@ -185,8 +185,47 @@ public class QuestionService {
 		}
 		Long userId = userRepository.getId(principal.getName());
 		Collection<Question> knownQuestions =
-				questionsRepository
+				questionRepository
 						.findQuestionsByUsersKnownThisQuestion(userId);
 		return new ResponseEntity<>(knownQuestions, HttpStatus.OK);
+	}
+
+	public ResponseEntity<Collection<Question>> getQuestionsByTagId(Long tagId) {
+		Optional<Tag> optionalTag = tagRepository.findById(tagId);
+		if (optionalTag.isPresent()) {
+			List<Question> questionList = questionRepository.findQuestionsByTagsContaining(optionalTag.get());
+			return new ResponseEntity<>(questionList, HttpStatus.OK);
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+	}
+
+	public ResponseEntity<List<Question>> getRandomByTag(Principal principal, Long tag_id) {
+		Page<Question> questionPage;
+		Tag tag;
+		Optional<Tag> optionalTag = tagRepository.findById(tag_id);
+		if (optionalTag.isPresent()) {
+			tag = optionalTag.get();
+		} else {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		if (principal != null && getCurrentUserKnownQuestionIds().size() > 0) {
+			Long quantity = questionRepository
+					.countQuestionsByTagsContainingAndAcceptedTrueAndIdNotIn(
+							tag, getCurrentUserKnownQuestionIds());
+			int index = (int) (Math.random() * quantity);
+			questionPage = questionRepository.findQuestionsByIdNotInAndAcceptedTrueAndTagsContaining(
+					getCurrentUserKnownQuestionIds(), tag, PageRequest.of(index, 1, Sort.unsorted()));
+		} else {
+			Long quantity = questionRepository.countQuestionByTagsContainingAndAcceptedTrue(tag);
+			int index = (int) (Math.random() * quantity);
+			questionPage = questionRepository
+					.findAllByTagsContainingAndAcceptedTrue(tag, PageRequest.of(index, 1, Sort.unsorted()));
+		}
+
+		if (questionPage.hasContent()) {
+			return new ResponseEntity<>(Collections.singletonList(questionPage.getContent().get(0)), HttpStatus.OK);
+		}
+		return new ResponseEntity<>(new ArrayList<>(), HttpStatus.OK);
 	}
 }
